Improvements on model
