+++ dto/priority.go
@@ -1,0 +1,20 @@
+package dto
+
+type Priorities struct {
+	P1  *float64 `json:"_1"`
+	P2  *float64 `json:"_2"`
+	P3  *float64 `json:"_3"`
+	P4  *float64 `json:"_4"`
+	P5  *float64 `json:"_5"`
+	P6  *float64 `json:"_6"`
+	P7  *float64 `json:"_7"`
+	P8  *float64 `json:"_8"`
+	P9  *float64 `json:"_9"`
+	P10 *float64 `json:"_10"`
+	P11 *float64 `json:"_11"`
+	P12 *float64 `json:"_12"`
+	P13 *float64 `json:"_13"`
+	P14 *float64 `json:"_14"`
+	P15 *float64 `json:"_15"`
+	P16 *float64 `json:"_16"`
+}
