+++ go.mod
@@ -3,14 +3,18 @@ module github.com/NubeIO/nubeio-rubix-lib-models-go
 go 1.18
 
 require (
-	gorm.io/datatypes v1.0.6
 	github.com/NubeIO/lib-systemctl-go v0.3.1
+	github.com/shopspring/decimal v1.2.0
+	gorm.io/datatypes v1.0.6
 )
 
 require (
+	github.com/NubeIO/lib-files v0.1.1 // indirect
 	github.com/go-sql-driver/mysql v1.6.0 // indirect
 	github.com/jinzhu/inflection v1.0.0 // indirect
 	github.com/jinzhu/now v1.1.4 // indirect
+	github.com/sirupsen/logrus v1.8.1 // indirect
+	golang.org/x/sys v0.0.0-20211216021012-1d35b9e2eb4e // indirect
 	gorm.io/driver/mysql v1.3.2 // indirect
 	gorm.io/gorm v1.23.2 // indirect
 )
