+++ model/alert.go
@@ -26,9 +26,3 @@ type Alert struct {
 	Teams        []*AlertTeam             `json:"teams,omitempty" gorm:"constraint:OnDelete:CASCADE"`
 	Transactions []*AlertTransaction      `json:"transactions,omitempty" gorm:"constraint:OnDelete:CASCADE"`
 }
-
-type AlertMetaTag struct {
-	AlertUUID string `json:"alert_uuid,omitempty" gorm:"type:varchar(255) references alerts;not null;default:null;primaryKey"`
-	Key       string `json:"key,omitempty" gorm:"primaryKey"`
-	Value     string `json:"value,omitempty"`
-}
