--- model/login.go
@@ -1,11 +1,0 @@
-package model
-
-type LoginBody struct {
-	Username string `json:"username"`
-	Password string `json:"password"`
-}
-
-type TokenResponse struct {
-	AccessToken string `json:"access_token"`
-	TokenType   string `json:"token_type"`
-}
