+++ model/point.go
@@ -7,22 +7,6 @@ import (
 	"time"
 )
 
-// TimeOverride TODO add in later
-// TimeOverride where a point value can be overridden for a duration of time
-type TimeOverride struct {
-	PointUUID string `json:"point_uuid" gorm:"references points;not null;default:null;primaryKey"`
-	StartDate string `json:"start_date"` // START at 25:11:2021:13:00
-	EndDate   string `json:"end_date"`   // START at 25:11:2021:13:30
-	Value     string `json:"value"`
-	Priority  string `json:"priority"`
-}
-
-// MathOperation same as in lora and point-server TODO add in later
-type MathOperation struct {
-	Calc string // x + 1
-	X    float64
-}
-
 // Point table
 type Point struct {
 	CommonUUID
@@ -96,113 +80,6 @@ type Point struct {
 	Config               datatypes.JSON `json:"config"`
 }
 
-type PointMetaTag struct {
-	PointUUID string `json:"point_uuid,omitempty" gorm:"type:varchar(255) references points;not null;default:null;primaryKey"`
-	Key       string `json:"key,omitempty" gorm:"primaryKey"`
-	Value     string `json:"value,omitempty"`
-}
-
-type Priorities struct {
-	P1  *float64 `json:"_1"`
-	P2  *float64 `json:"_2"`
-	P3  *float64 `json:"_3"`
-	P4  *float64 `json:"_4"`
-	P5  *float64 `json:"_5"`
-	P6  *float64 `json:"_6"`
-	P7  *float64 `json:"_7"`
-	P8  *float64 `json:"_8"`
-	P9  *float64 `json:"_9"`
-	P10 *float64 `json:"_10"`
-	P11 *float64 `json:"_11"`
-	P12 *float64 `json:"_12"`
-	P13 *float64 `json:"_13"`
-	P14 *float64 `json:"_14"`
-	P15 *float64 `json:"_15"`
-	P16 *float64 `json:"_16"`
-}
-
-type Priority struct {
-	PointUUID string   `json:"point_uuid,omitempty" gorm:"references points;not null;default:null;primaryKey"`
-	P1        *float64 `json:"_1"`
-	P2        *float64 `json:"_2"`
-	P3        *float64 `json:"_3"`
-	P4        *float64 `json:"_4"`
-	P5        *float64 `json:"_5"`
-	P6        *float64 `json:"_6"`
-	P7        *float64 `json:"_7"`
-	P8        *float64 `json:"_8"`
-	P9        *float64 `json:"_9"`
-	P10       *float64 `json:"_10"`
-	P11       *float64 `json:"_11"`
-	P12       *float64 `json:"_12"`
-	P13       *float64 `json:"_13"`
-	P14       *float64 `json:"_14"`
-	P15       *float64 `json:"_15"`
-	P16       *float64 `json:"_16"`
-}
-
-type PointWriter struct {
-	Priority     *map[string]*float64 `json:"priority"`
-	PresentValue *float64             `json:"present_value"`
-	ForceWrite   bool                 `json:"force_write"`
-	Message      string               `json:"message"`
-	Fault        bool                 `json:"fault"`
-	PollState    datatype.PointState  `json:"poll_state"`
-	Timestamp    *time.Time           `json:"timestamp"`
-}
-
-func (p *Priority) GetHighestPriorityValue() *float64 {
-	if p.P1 != nil {
-		return p.P1
-	}
-	if p.P2 != nil {
-		return p.P2
-	}
-	if p.P3 != nil {
-		return p.P3
-	}
-	if p.P4 != nil {
-		return p.P4
-	}
-	if p.P5 != nil {
-		return p.P5
-	}
-	if p.P6 != nil {
-		return p.P6
-	}
-	if p.P7 != nil {
-		return p.P7
-	}
-	if p.P8 != nil {
-		return p.P8
-	}
-	if p.P9 != nil {
-		return p.P9
-	}
-	if p.P10 != nil {
-		return p.P10
-	}
-	if p.P11 != nil {
-		return p.P11
-	}
-	if p.P12 != nil {
-		return p.P12
-	}
-	if p.P13 != nil {
-		return p.P13
-	}
-	if p.P14 != nil {
-		return p.P14
-	}
-	if p.P15 != nil {
-		return p.P15
-	}
-	if p.P16 != nil {
-		return p.P16
-	}
-	return nil
-}
-
 // GetPntType find the type of point as in voltage_dc
 func GetPntType(strut interface{}, ioType string) (out string, err error) {
 	val := reflect.ValueOf(strut).Elem()
@@ -214,60 +91,4 @@ func GetPntType(strut interface{}, ioType string) (out string, err error) {
 		}
 	}
 	return
-}
-
-type PointWithParent struct {
-	UUID        string `json:"uuid"`
-	Name        string `json:"name"`
-	DeviceUUID  string `json:"device_uuid"`
-	DeviceName  string `json:"device_name"`
-	NetworkUUID string `json:"network_uuid"`
-	NetworkName string `json:"network_name"`
-}
-
-type PointHistoryInterval struct {
-	UUID            string   `json:"uuid"`
-	HistoryInterval *int     `json:"history_interval,omitempty"`
-	Timestamp       string   `json:"timestamp,omitempty"`
-	PresentValue    *float64 `json:"present_value,omitempty"`
-}
-
-type PointForPostgresSync struct {
-	UUID         string `json:"uuid"`
-	Name         string `json:"name"`
-	DeviceUUID   string `json:"device_uuid,omitempty"`
-	DeviceName   string `json:"device_name,omitempty"`
-	NetworkUUID  string `json:"network_uuid"`
-	NetworkName  string `json:"network_name"`
-	GlobalUUID   string `json:"global_uuid"`
-	HostUUID     string `json:"host_uuid"`
-	HostName     string `json:"host_name"`
-	GroupUUID    string `json:"group_uuid"`
-	GroupName    string `json:"group_name"`
-	LocationUUID string `json:"location_uuid"`
-	LocationName string `json:"location_name"`
-}
-
-type PointTagForPostgresSync struct {
-	HostUUID  string `json:"host_uuid"`
-	PointUUID string `json:"point_uuid"`
-	Tag       string `json:"tag"`
-}
-
-type PointMetaTagForPostgresSync struct {
-	HostUUID  string `json:"host_uuid,omitempty"`
-	PointUUID string `json:"point_uuid,omitempty"`
-	Key       string `json:"key,omitempty"`
-	Value     string `json:"value,omitempty"`
-}
-
-type UpdatePointOpts struct {
-	WriteValue bool
-}
-
-type PointWriteResponse struct {
-	Point                Point `json:"point"`
-	IsPresentValueChange bool  `json:"is_present_value_change"`
-	IsWriteValueChange   bool  `json:"is_write_value_change"`
-	IsPriorityChanged    bool  `json:"is_priority_changed"`
 }
