+++ model/priority.go
@@ -1,0 +1,73 @@
+package model
+
+type Priority struct {
+	PointUUID string   `json:"point_uuid,omitempty" gorm:"references points;not null;default:null;primaryKey"`
+	P1        *float64 `json:"_1"`
+	P2        *float64 `json:"_2"`
+	P3        *float64 `json:"_3"`
+	P4        *float64 `json:"_4"`
+	P5        *float64 `json:"_5"`
+	P6        *float64 `json:"_6"`
+	P7        *float64 `json:"_7"`
+	P8        *float64 `json:"_8"`
+	P9        *float64 `json:"_9"`
+	P10       *float64 `json:"_10"`
+	P11       *float64 `json:"_11"`
+	P12       *float64 `json:"_12"`
+	P13       *float64 `json:"_13"`
+	P14       *float64 `json:"_14"`
+	P15       *float64 `json:"_15"`
+	P16       *float64 `json:"_16"`
+}
+
+func (p *Priority) GetHighestPriorityValue() *float64 {
+	if p.P1 != nil {
+		return p.P1
+	}
+	if p.P2 != nil {
+		return p.P2
+	}
+	if p.P3 != nil {
+		return p.P3
+	}
+	if p.P4 != nil {
+		return p.P4
+	}
+	if p.P5 != nil {
+		return p.P5
+	}
+	if p.P6 != nil {
+		return p.P6
+	}
+	if p.P7 != nil {
+		return p.P7
+	}
+	if p.P8 != nil {
+		return p.P8
+	}
+	if p.P9 != nil {
+		return p.P9
+	}
+	if p.P10 != nil {
+		return p.P10
+	}
+	if p.P11 != nil {
+		return p.P11
+	}
+	if p.P12 != nil {
+		return p.P12
+	}
+	if p.P13 != nil {
+		return p.P13
+	}
+	if p.P14 != nil {
+		return p.P14
+	}
+	if p.P15 != nil {
+		return p.P15
+	}
+	if p.P16 != nil {
+		return p.P16
+	}
+	return nil
+}
